Merges morse_led_node and detect_node, adding Morse as a detection type. Implementation can probably be made cleaner once I can tell it works.
